backend/requirements.txt
beautifulsoup4
ddddocr==1.5.6
django==5.1.3
djangorestframework==3.15.0
eventlet>=0.24.1
faker
gunicorn
httpx
pillow
psycopg[binary]
pyjwt
python-dotenv==1.0.0
scrypt